Moved timetable tab popups to separate file + tidied up some of the logic for timetable tabs
